Changed file locations & Added .gitignore
