+++ aruco/launch/aruco.launch
@@ -1,0 +1,3 @@
+<launch>
+	<node name="aruco" pkg="aruco" type="estimator.py" output="screen"></node>
+</launch>
