+++ aruco/scripts/aruco_ids.py
@@ -1,0 +1,39 @@
+id_72 = {
+    "x_position": -30,
+    "y_position": 100
+}
+
+id_75 = {
+    "x_position": 30,
+    "y_position": 100
+}
+
+id_78 = {
+    "x_position": -30,
+    "y_position": 190
+}
+
+id_88 = {
+    "x_position": 30,
+    "y_position": 190
+}
+
+id_91 = {
+    "x_position": -30,
+    "y_position": 280
+}
+
+id_97 = {
+    "x_position": 30,
+    "y_position": 280
+}
+
+id_98 = {
+    "x_position": -30,
+    "y_position": 370
+}
+
+id_104 = {
+    "x_position": 30,
+    "y_position": 370
+}
