+++ aruco/scripts/cameraDistortion.txt
@@ -1,0 +1,1 @@
+8.909644199915237039e-02,-4.709586824483883993e-01,-2.324722018434232859e-03,-3.513643215420087364e-04,7.751848670595281598e-01
