+++ aruco/scripts/cameraMatrix.txt
@@ -1,0 +1,3 @@
+6.365712249823847060e+02,0.000000000000000000e+00,3.137140925022698639e+02
+0.000000000000000000e+00,6.371694838205387441e+02,2.303806556083854105e+02
+0.000000000000000000e+00,0.000000000000000000e+00,1.000000000000000000e+00
