--- src/aruco/scripts/newfile.log
@@ -1,1 +1,0 @@
-2022-10-26 18:44:48,282 tm._add: /position, geometry_msgs/PoseStamped, pub
